% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeersonSasorov.R
\name{burnout_prob_MS}
\alias{burnout_prob_MS}
\title{Burnout probability based on Meerson and Sasorov (2009)}
\usage{
burnout_prob_MS(
  R0,
  epsilon,
  N = 10^6,
  subdivisions = 1000L,
  debug = FALSE,
  tiny = 0,
  ...
)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}

\item{N}{population size}

\item{subdivisions}{the maximum number of subintervals.}

\item{debug}{print debugging info?}

\item{tiny}{avoid integration limits by this amount}

\item{...}{additional arguments pass to
\code{\link[stats]{integrate}}}
}
\description{
Burnout probability based on Meerson and Sasorov (2009)
}
\details{
See equation (9) of Ballard et al (2016)

The derivation assumes \eqn{N S_0 \gg 1}, where \eqn{S_0} is
defined in equation (9) of Ballard et al (2016).
}
\references{
\insertRef{Meerson2009}{burnout}
}
\seealso{
\code{\link{P1_prob_MS}}
}
