% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanH.R
\name{burnout_prob_vanH}
\alias{burnout_prob_vanH}
\title{Burnout probability based on van Herwaarden (1997)}
\usage{
burnout_prob_vanH(
  R0,
  epsilon,
  N = 10^6,
  subdivisions = 1000L,
  tiny = 0,
  debug = FALSE,
  persist = FALSE,
  ...
)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}

\item{N}{population size}

\item{subdivisions}{the maximum number of subintervals.}

\item{tiny}{amount which to avoid integration limits (where
integrand blow up)}

\item{debug}{print debugging info?}

\item{persist}{return persistence probability rather than burnout probability}

\item{...}{additional arguments pass to
\code{\link[stats]{integrate}}}
}
\description{
Burnout probability based on van Herwaarden (1997)
}
\details{
See equation (8) of Ballard et al (2016)
}
\references{
\insertRef{Ballard2016}{burnout}

\insertRef{vanH1997}{burnout}
}
\seealso{
\code{\link{P1_prob_vanH}}
}
