% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_funs}
\alias{compare_funs}
\title{Compare functions of \eqn{{\cal R}_0} and \eqn{\varepsilon}}
\usage{
compare_funs(
  fun1 = q_exact,
  fun2 = q_approx,
  R0 = NULL,
  Rmin = 1.001,
  Rmax = 64,
  nR0 = 101,
  epsilon = 10^(-(4:1)),
  nepsilon = length(epsilon),
  N = NULL,
  ...,
  show.progress = FALSE
)
}
\arguments{
\item{fun1, fun2}{functions of \code{R0} and \code{epsilon} to be
compared}

\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{Rmin, Rmax}{minimum and maximum values of \eqn{{\cal R}_0}}

\item{nR0, nepsilon}{number of values of \code{R0} and
\code{epsilon}}

\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}

\item{N}{population size}

\item{...}{additional parameters passed to \code{fun1} and
\code{fun2}}

\item{show.progress}{logical: if \code{TRUE} then spew each
element of the data frame to \code{stdout} as it is computed}
}
\value{
data frame with \eqn{\code{nR0*nepsilon}} rows
}
\description{
Compare functions of \eqn{{\cal R}_0} and \eqn{\varepsilon}
}
\details{
The default \code{R0} value of \code{NULL} yields a
sensible sprinkling of \code{nR0} \eqn{{\cal R}_0} values.
Typical uses are to compare exact vs approximate
calculations of Kendall's \eqn{q}, or different approximations
of \eqn{x_{\rm in}}.
}
\examples{
(cf <- compare_funs(nR0=101))
}
\seealso{
\code{\link{q_exact}}, \code{\link{q_approx}},
\code{\link{x_in}}
}
