% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_ij.R
\name{Ein}
\alias{Ein}
\title{Ein special function}
\usage{
Ein(z)
}
\arguments{
\item{z}{real or complex number}
}
\description{
\eqn{Ein(z) = E_1(z) + \log(z) + \gamma}
}
\seealso{
\code{\link[expint]{expint_E1}}
}
