% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_in.R
\name{eqm_prev}
\alias{eqm_prev}
\title{Equilibrium prevalence \eqn{y^\star}}
\usage{
eqm_prev(R0, epsilon, eta.i = 0, eta.a = 0)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}
}
\description{
Equilibrium prevalence \eqn{y^\star}
}
\details{
If there is no effective waning from antigenic evolution,
i.e., \eqn{\alpha=0}, then
\deqn{
	y^\star = \frac{\varepsilon+\eta}{1+\eta}
      \big( 1 - \frac{1}{{\mathcal R}_0} \big)
}
}
