% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_in.R
\name{final_size}
\alias{final_size}
\title{Standard final size}
\usage{
final_size(R0)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}
}
\description{
for a newly invading infectious disease (so \eqn{I_0\to0}, \eqn{S_0\to1}).
\deqn{
  Z = 1 + \frac{1}{{\mathcal R}_{0}} W_{0}\left(-{\mathcal R}_{0} e^{-{\mathcal R}_{0}}\right)
}
}
\examples{
final_size(R0 = 2)
}
