% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fizzle.R
\name{fizzle_prob}
\alias{fizzle_prob}
\title{Fizzle probability}
\usage{
fizzle_prob(R0, k = 1)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{k}{initial number of infected individuals (\eqn{k})}
}
\value{
real number between 0 and 1
}
\description{
Probability that a disease will go extinct before causing a
substantial outbreak.
}
\details{
The fizzle probability is

\deqn{1 - p_k =
  \left\{
  \begin{array}{rl}
     1\,, & 0 \le {\cal R}_0 \le 1, \\
     \left(\frac{1}{{\cal R}_0}\right)^{k} \,, & 1\le{\cal R}_0.
  \end{array}
  \right.
}
}
\examples{
fizzle_prob(R0=c(1/2,2))
fizzle_prob(R0=2, k=2)
fizzle_prob(R0=c(2,4,8))
fizzle_prob(R0=2, k=c(1,2))  # FIX: fails with vector k

}
\seealso{
\code{\link{fizzle_prob}}
}
