% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fizzle.R
\name{not_fizzle_prob}
\alias{not_fizzle_prob}
\title{Not fizzle probability}
\usage{
not_fizzle_prob(R0, k = 1)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{k}{initial number of infected individuals (\eqn{k})}
}
\value{
real number between 0 and 1
}
\description{
Probability that a disease will \emph{not} go extinct before
causing a substantial outbreak.
}
\details{
The probability of not fizzling is

\deqn{p_k =
  \left\{
  \begin{array}{rl}
     0\,, & 0 \le {\cal R}_0 \le 1, \\
     1 - \left(\frac{1}{{\cal R}_0}\right)^{k} \,, & 1\le{\cal R}_0.
  \end{array}
  \right.
}
}
\examples{
not_fizzle_prob(R0=c(1/2,2))
not_fizzle_prob(R0=2, k=2)
not_fizzle_prob(R0=c(2,4,8))
not_fizzle_prob(R0=2, k=c(1,2))  # FIX: fails with vector k

}
\seealso{
\code{\link{fizzle_prob}}
}
