% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P1.R
\name{P1_prob}
\alias{P1_prob}
\title{Persistence probability}
\usage{
P1_prob(
  R0,
  epsilon,
  eta.i = 0,
  eta.a = 0,
  k = 1,
  N = 10^6,
  xin = x_in(R0 = R0, epsilon = epsilon, eta.i = eta.i, eta.a = eta.a),
  q_fun = q_approx,
  ystar = eqm_prev(R0 = R0, epsilon = epsilon, eta.i = eta.i, eta.a = eta.a)
)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}

\item{k}{initial number of infected individuals (\eqn{k})}

\item{N}{population size}

\item{xin}{by default set via the approximation coded in \code{\link{x_in}}}

\item{q_fun}{function to compute Kendall's \eqn{q}}

\item{ystar}{prevalence at boundary layer (equilibrium prevalence
by default)}
}
\value{
real number between 0 and 1
}
\description{
Probability of surviving the trough after the first epidemic.
}
\details{
If \eqn{p_k} is the probability of \emph{not}
\code{\link[=fizzle_prob]{fizzling}} if \eqn{k} individuals are
initially infected, then the probability of burning out
conditional on not fizzling is

\deqn{ p_k q^{Ny^\star} \,,}

where \eqn{q} is Kendall's \eqn{q}, the expected total population
size is \eqn{N}, and the equilibrium prevalence is

\deqn{y^\star = \varepsilon\Big(1 - \frac{1}{{\cal R}_0}\Big) \,.}

Hence, the probability of \emph{either} fizzling before a first
major outbreak \emph{or} burning out after a first major outbreak
is

\deqn{
     (1-p_k) + p_k q^{Ny^\star} =
     1 - p_k\big( 1 - q^{Ny^\star} \big) \,.
}

Consequently, the probability of \emph{persisting} beyond the
first epidemic is

\deqn{
    {{\cal P}_1}({\cal R}_0,\varepsilon,k,N) =
      p_k\big( 1 - q^{Ny^\star} \big) \,.
}
}
\examples{
P1_prob(R0=2, epsilon=0.01)
P1_prob(R0=2, epsilon=0)
P1_prob(R0=c(2,4), epsilon=0.01)
P1_prob(R0=2, epsilon=c(0.01, 0.1))
curve(P1_prob(x,epsilon=0.001), from=1.01, to=2, las=1, n=1001, ylim=c(0,1))
curve(P1_prob(x,epsilon=0.01), from=1.01, to=2, las=1, add=TRUE, col="magenta", n=1001)
curve(P1_prob(x,epsilon=0.1), from=1.01, to=2, las=1, add=TRUE, col="cyan", n=1001)
xvals <- exp(seq(log(1.01), log(64), length=1001))
P1vals <- P1_prob(xvals,epsilon=0.01)
## results are complex:
max(Im(P1vals))
P1vals <- Re(P1vals)
plot(xvals, P1vals, las=1, col="blue", type="l", lwd=2, log="x")
cat("max(P1) = ", max(P1vals), "\n")

}
\references{
\insertRef{Kendall1948b}{burnout}
}
\seealso{
\code{\link{fizzle_prob}}, \code{\link{x_in}},
\code{\link{llig}}, \code{\link{q_approx}}, \code{\link{q_exact}}
}
