% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanH.R
\name{P1_prob_vanH}
\alias{P1_prob_vanH}
\title{Persistence probability via van Herwaarden (1997) approximation}
\usage{
P1_prob_vanH(
  R0,
  epsilon,
  k = 1,
  N = 10^6,
  subdivisions = 1000L,
  persist = FALSE,
  tiny = 0,
  ...
)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}

\item{k}{initial number of infected individuals (\eqn{k})}

\item{N}{population size}

\item{subdivisions}{the maximum number of subintervals.}

\item{persist}{return persistence probability rather than burnout probability}

\item{tiny}{avoid integration limits by this amount}

\item{...}{additional arguments pass to
\code{\link[stats]{integrate}}}
}
\description{
Persistence probability via van Herwaarden (1997) approximation
}
\references{
\insertRef{Ballard2016}{burnout}

\insertRef{vanH1997}{burnout}
}
\seealso{
\code{\link{burnout_prob_vanH}}, \code{\link{P1_prob_MS}},
\code{\link{P1_prob}}
}
