% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_prev.R
\name{peak_prev_nvd}
\alias{peak_prev_nvd}
\title{Peak prevalence}
\usage{
peak_prev_nvd(R0, xi = 1, yi = 0, ...)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{xi}{initial susceptible proportion \eqn{x_{\rm i}}}

\item{yi}{initial infective proportion \eqn{y_{\rm i}}}

\item{...}{(ignored) for compatibility
with the more accurate \code{\link{peak_prev}} function
that takes additional arguments}
}
\value{
real number between 0 and 1
}
\description{
Exact peak prevalence in the SIR model with \emph{no vital dynamics} (nvd).
}
\details{
\deqn{
  y_{\rm max} = 1 -
    \frac{1}{{\cal R}_{0}}\big(1 +\ln{{\cal R}_{0}}\big) \,.
}
}
\seealso{
\code{\link{peak_prev}}, \code{\link{x_in}}
}
