% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.compare_funs}
\alias{plot.compare_funs}
\title{Plot comparison of two functions.}
\usage{
\method{plot}{compare_funs}(x, show.plots = TRUE, ...)
}
\arguments{
\item{x}{\code{\link{compare_funs}} object, i.e., a
\code{\link{data.frame}} with particular structure}

\item{show.plots}{logical: if \code{FALSE} then return plots in a
list without displaying them}

\item{...}{additional parameters passed to \code{\link{plot}}}
}
\description{
\code{\link{plot}} method for \code{\link{compare_funs}} objects.
Typical use: exact vs approximate \eqn{q}.
}
\seealso{
\code{\link{compare_funs}}, \code{\link{q_exact}},
\code{\link{q_approx}}, \code{\link{x_in}}
}
