% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_in.R
\name{plot_all_x_in}
\alias{plot_all_x_in}
\title{Plot all approximations to \eqn{x_{\rm in}}}
\usage{
plot_all_x_in(
  Rmin = 1.0001,
  Rmax = 20,
  R0 = seq(Rmin, Rmax, length = 1001),
  epsilon = c(0.01, 0.02, 0.03),
  xvar = "R0",
  col = 1:length(epsilon),
  lwd = 2,
  log = "x",
  xlim = if (xvar == "R0") c(1, Rmax) else c(0, 1),
  ylim = if (xvar == "R0") c(0, 1) else c(0, max(1/R0)),
  ...
)
}
\arguments{
\item{Rmin, Rmax}{minimum and maximum values of \eqn{{\cal R}_0}}

\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}

\item{xvar}{string indicating whether \code{R0} or \code{epsilon}
is the desired independent variable}

\item{col, lwd, xlim, ylim, log, ...}{see \code{\link{graphical parameters}}}
}
\description{
Plot all approximations to \eqn{x_{\rm in}}
}
\seealso{
\code{\link{plot_x_in}}, \code{\link{x_in}},
\code{\link{x_in_exact}}
}
