% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fizzle.R
\name{plot_fizzle_time}
\alias{plot_fizzle_time}
\title{Plot fizzle time}
\usage{
plot_fizzle_time(
  R0 = exp(seq(log(1.001), log(2), length = 1001)),
  k = 1:3,
  delta = 0.1,
  col = 1:length(k),
  lwd = 2,
  log = "x",
  ...
)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{k}{initial number of infected individuals (\eqn{k})}

\item{delta}{probability threshold}

\item{col, lwd, log, ...}{see \code{\link{graphical parameters}}}
}
\description{
Plot fizzle time
}
\examples{
plot_fizzle_time(delta = 0.01)

}
\seealso{
\code{\link{fizzle_time}}
}
