% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_P1}
\alias{plot_P1}
\title{Plot persistence probability}
\usage{
plot_P1(
  epsilon = 0.01,
  N = 10^6,
  Rmin = 1.001,
  Rmax = 64,
  P1_fun = P1_prob,
  add = FALSE,
  col = "black",
  lwd = 2,
  log = "x",
  show.N = TRUE,
  show.epsilon = FALSE,
  ...
)
}
\arguments{
\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}

\item{N}{population size}

\item{Rmin, Rmax}{minimum and maximum value of \eqn{{\cal R}_0}}

\item{P1_fun}{function that computes \eqn{{\cal P}_1}}

\item{add}{if \code{TRUE}, add to existing plot}

\item{col, lwd, log}{standard \link{graphical parameters}}

\item{show.N, show.epsilon}{if \code{TRUE}, display parameter value in legend}

\item{...}{additional parameters passed to \code{\link{plot}}}
}
\description{
Plot persistence probability
}
\examples{
op <- par(mfrow = c(2,2))
plot_P1(epsilon=0.01, N=10^4)
plot_P1(epsilon=0.01, N=10^5)
plot_P1(epsilon=0.01, N=10^6)
plot_P1(epsilon=0.01, N=10^7)
par(mfrow = op)

}
\seealso{
P1_prob
}
