% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_herd.R
\name{plot_t_herd}
\alias{plot_t_herd}
\title{Plot herd immunity duration}
\usage{
plot_t_herd(
  R0 = exp(seq(log(1.001), log(5), length = 1001)),
  epsilon = c(0.01, 0.02, 0.03),
  x_in_fun = x_in_exact,
  col = 1:length(epsilon),
  lwd = 2,
  lty = "solid",
  ylim = c(0, 60),
  xlab = expression(R[0]),
  ylab = expression(t[H]),
  main = latex2exp::TeX("Duration of herd immunity $t_{H}$"),
  show.legend = TRUE,
  add = FALSE,
  ...
)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}

\item{x_in_fun}{function to use to estimate \eqn{x_{\rm in}}}

\item{col, lwd, ylim, lty, main, xlab, ylab, ...}{see \code{\link{graphical parameters}}}

\item{show.legend}{if \code{TRUE} then show legend}

\item{add}{if \code{TRUE} then add to existing plot}
}
\description{
Plot herd immunity duration
}
\details{
It is best to use \code{\link{x_in_exact}} when
calculating \code{t_herd}.
}
\examples{
## plot tH with x_in as dotted under tH with x_in_exact as solid:
plot_t_herd(x_in_fun = x_in, lty="dotted")
plot_t_herd(add=TRUE)

}
\seealso{
\code{\link{t_herd}}
}
