% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_in.R
\name{plot_x_in}
\alias{plot_x_in}
\title{Plot \eqn{x_{\rm in}}}
\usage{
plot_x_in(
  Rmin = 1.0001,
  Rmax = 20,
  R0 = seq(Rmin, Rmax, length = 1001),
  epsilon = c(0.01, 0.02, 0.03),
  xvar = "R0",
  xin_fun = x_in,
  col = 1:length(epsilon),
  lwd = 2,
  lty = "solid",
  log = "x",
  main = "Susceptible proportion at\\nentry to boundary layer",
  xlim = if (xvar == "R0") c(1, Rmax) else c(0, 1),
  ylim = if (xvar == "R0") c(0, 1) else c(0, max(1/R0)),
  add = FALSE,
  ...
)
}
\arguments{
\item{Rmin, Rmax}{minimum and maximum values of \eqn{{\cal R}_0}}

\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}

\item{xvar}{string indicating whether \code{R0} or \code{epsilon}
is the desired independent variable}

\item{xin_fun}{function of \code{R0} and \code{epsilon} that
returns susceptible proportion at entry to boundary layer}

\item{col, log, lwd, xlim, ylim, lty, main, ...}{see \code{\link{graphical parameters}}}

\item{add}{if \code{TRUE} then add to existing plot}
}
\description{
Plot \eqn{x_{\rm in}}
}
\examples{
\dontrun{
op <- par(mfrow = c(1,2))
plot_x_in(epsilon = 0.01)
plot_x_in(epsilon = 0.1)
par(mfrow = op)
op <- par(mfrow = c(1,3))
plot_x_in(R0 = 2, epsilon = seq(0,1,length=101), xvar="epsilon", log="")
plot_x_in(R0 = 5, epsilon = seq(0,1,length=101), xvar="epsilon", log="")
plot_x_in(R0 = 20, epsilon = seq(0,1,length=101), xvar="epsilon", log="")
par(mfrow = op)
}

}
\seealso{
\code{\link{x_in}}, \code{\link{x_in_exact}}
}
