% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pm.R
\name{Pm_prob}
\alias{Pm_prob}
\title{Persistence probability after \eqn{m}'th epidemic wave}
\usage{
Pm_prob(
  R0,
  epsilon,
  m,
  k = 1,
  N = 10^6,
  xin = x_in(R0, epsilon),
  q_fun = q_approx,
  ystar = epsilon * (1 - 1/R0)
)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}

\item{m}{epidemic wave number}

\item{k}{initial number of infected individuals (\eqn{k})}

\item{N}{population size}

\item{xin}{by default set via the approximation coded in \code{\link{x_in}}}

\item{q_fun}{function to compute Kendall's \eqn{q}}

\item{ystar}{prevalence at boundary layer (equilibrium prevalence
by default)}
}
\value{
real number between 0 and 1
}
\description{
Probability of surviving the trough after the \eqn{m}'th epidemic.
}
\details{
\deqn{
    {{\cal P}_m}({\cal R}_0,\varepsilon,k,N) =
      p_k\prod_{j=1}^m\big( 1 - q(x_{{\rm in},j})^{Ny^\star} \big) \,.
}
}
\note{
This function is recursive and ultimate calls
\code{\link{P1_prob}}.
}
\examples{
\dontrun{
Pm_prob(R0=2, epsilon=0.01, m=2)
Pm_prob(R0=2, epsilon=0, m=2)
Pm_prob(R0=c(2,4), epsilon=0.01, m=2)
Pm_prob(R0=2, epsilon=c(0.01, 0.1), m=2)
curve(Pm_prob(x,epsilon=0.001, m=2), from=1.01, to=2, las=1, n=1001, ylim=c(0,1))
curve(Pm_prob(x,epsilon=0.01, m=2), from=1.01, to=2, las=1, add=TRUE, col="magenta", n=1001)
curve(Pm_prob(x,epsilon=0.1, m=2), from=1.01, to=2, las=1, add=TRUE, col="cyan", n=1001)
xvals <- exp(seq(log(1.01), log(64), length=1001))
Pmvals <- Pm_prob(xvals,epsilon=0.01, m=2)
## results are complex:
max(Im(Pmvals))
Pmvals <- Re(Pmvals)
plot(xvals, Pmvals, las=1, col="blue", type="l", lwd=2, log="x")
cat("max(Pm) = ", max(Pmvals), "\n")
}

}
\references{
\insertRef{Kendall1948b}{burnout}
}
\seealso{
\code{\link{P1_prob}}, \code{\link{fizzle_prob}},
\code{\link{x_in}}, \code{\link{llig}},
\code{\link{q_approx}}, \code{\link{q_exact}}
}
