% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.R
\name{q_exact}
\alias{q_exact}
\title{Kendall's \eqn{q} (exact expression)}
\usage{
q_exact(
  R0,
  epsilon,
  eta.i = 0,
  eta.a = 0,
  xin = x_in(R0, epsilon, eta.i, eta.a)
)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}

\item{xin}{by default set via the approximation coded in \code{\link{x_in}}}
}
\value{
real number between 0 and 1
}
\description{
Probability of eventual extinction in a birth-death process
starting from one individual.
}
\details{
The theorem of \insertCite{Kendall1948b;textual}{burnout} applied
to the stochastic SIR model with one infected individual yields

\deqn{
	q({\cal R}_0,\varepsilon,x_{\rm in}) =
\left(1+\frac{\varepsilon}{e^{\frac{{\cal R}_{0}}{\varepsilon} (1-x_{\rm in})}
\left(\frac{{\cal R}_{0}}{\varepsilon}
(1-x_{\rm in})\right)^{-\frac{{\cal R}_{0}}{\varepsilon}\left(1-\frac{1}{{\cal R}_{0}}\right)}
{\cal g}\left(\frac{{\cal R}_{0}}{\varepsilon}\left(1-\frac{1}{{\cal R}_{0}}\right),\frac{{\cal R}_{0}}{\varepsilon}(1-x_{\rm in})\right)}\right)^{-1}.
}
}
\examples{
q_exact(R0=2, epsilon=0.01)
q_exact(R0=2, epsilon=0)
q_exact(R0=c(2,4), epsilon=0.01)
q_exact(R0=2, epsilon=c(0.01, 0.1))
curve(q_exact(x,epsilon=0.001), from=1.01, to=2, las=1, n=1001, ylim=c(0,1))
curve(q_exact(x,epsilon=0.01), from=1.01, to=2, las=1, add=TRUE, col="magenta", n=1001)
curve(q_exact(x,epsilon=0.1), from=1.01, to=2, las=1, add=TRUE, col="cyan", n=1001)

}
\references{
\insertRef{Kendall1948b}{burnout}
}
\seealso{
\code{\link{x_in}}, \code{\link{llig}}, \code{\link{q_approx}}
}
