% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_in.R
\name{R0_min_for_x_in}
\alias{R0_min_for_x_in}
\title{Minimum \eqn{{\mathcal R}_0} for which \eqn{x_{\rm in}}
approximation is valid}
\usage{
R0_min_for_x_in(epsilon)
}
\arguments{
\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}
}
\description{
Minimum \eqn{{\mathcal R}_0} for which \eqn{x_{\rm in}}
approximation is valid
}
\details{
\deqn{
	{\mathcal R}_0 >
    -\frac{W_{-1}\left(-(1-\varepsilon)
     e^{-(1-\varepsilon)}\right)}{1-\varepsilon}
}
}
