% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirias.R
\name{sirias_traj}
\alias{sirias_traj}
\title{SIRiaS trajectory computation via deSolve}
\usage{
sirias_traj(
  R0,
  epsilon,
  eta.i = 0,
  eta.a = 0,
  I0 = 1e-06,
  t_int = NULL,
  tmin = 20,
  maxit = 10,
  nt = 1000
)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}

\item{I0}{initial prevalence (proportion)}

\item{t_int}{integration time (default is max(tmin, 100/(R0-1)))}

\item{tmin}{minimum integration time}

\item{maxit}{max number of tries to double integration time}

\item{nt}{number of time steps}
}
\description{
SIRiaS trajectory computation via deSolve
}
