% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirias.R
\name{sirias_vf}
\alias{sirias_vf}
\title{Vector field for SIRiaS model}
\usage{
sirias_vf(tau, vars, parms)
}
\arguments{
\item{tau}{time in natural units}

\item{vars}{state variable list (\eqn{S}, \eqn{\log{I}}, \eqn{R})}

\item{parms}{parameter list}
}
\description{
Vector field for SIRiaS model
}
