% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_herd.R
\name{t_herd}
\alias{t_herd}
\title{Duration of herd immunity}
\usage{
t_herd(R0, epsilon, xin = x_in(R0, epsilon))
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}

\item{xin}{by default set via the approximation coded in \code{\link{x_in}}}
}
\description{
The time \eqn{t_{\rm H}} (after the first epidemic) from when the
proportion susceptible drops below the herd immunity threshold
(defined via \eqn{X(t_{\rm H}) = 1/{\cal R}_0}) until the
proportion susceptible next rises above the threshold.
}
\details{
\deqn{
	t_{\rm H} =  \frac{1}{\varepsilon}
    \ln\Big( \frac{1-x_{\rm in}}{1 - \frac{1}{{\cal R}_0}} \Big)
}
}
\examples{
\dontrun{
op <- par(mfrow = c(1,2))
plot_t_herd(col = c("darkred", "darkgreen", "darkblue"), main="")
epsilonseq <- seq(0,0.03,length=1001)
plot(epsilonseq, t_herd(R0=1.5,epsilonseq), type="l", lwd=2, bty="L",
     las=1, col="darkred", ylim=c(0,60),
     xlab = expression(epsilon), ylab = expression(t[H]))
lines(epsilonseq, t_herd(R0=3,epsilonseq), lwd=2, col="darkgreen")
lines(epsilonseq, t_herd(R0=4.5,epsilonseq), lwd=2, col="darkblue")
par(mfrow = op)
title(main=latex2exp::TeX("Duration of herd immunity $t_{H}$"))
}
}
