% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transparent_colour}
\alias{transparent_colour}
\title{Transparent colour}
\usage{
transparent_colour(col, alpha = 150/255)
}
\arguments{
\item{col}{as in \code{\link{col2rgb}}: vector of any of the three
kinds of R color specifications, i.e., either a colour name
(as listed by \code{colors()}), a hexadecimal string of the
form \code{"##rrggbb"} or \code{"##rrggbbaa"} (see
\code{\link{rgb}}), or a positive integer \code{i} meaning
\code{\link{palette}}\code{()[i]}.}

\item{alpha}{opacity level on scale from \code{0} to \code{1}
(i.e., \code{1} means completely opaque)}
}
\description{
Given a colour, change it to a transparent version
with a given \code{alpha} level
}
\seealso{
\code{\link{col2rgb}}, \code{\link{rgb}}
}
