% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_ij.R
\name{x_i}
\alias{x_i}
\title{Recursive function to compute \eqn{x_{{\rm i},j}}}
\usage{
x_i(R0, j)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{j}{epidemic wave index}
}
\description{
This uses the recurrence relation from Parsons and Earn (2023).
FIX: cite paper.
}
\seealso{
\code{\link{P1_prob}}
}
