% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_in.R
\name{x_in_cb}
\alias{x_in_cb}
\title{Susceptibles at boundary layer (corner/boundary layer approximation)}
\usage{
x_in_cb(R0, epsilon, peakprev_fun = peak_prev, ...)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}

\item{peakprev_fun}{function of \eqn{{\cal R}_0} and
\eqn{\varepsilon} to use to compute peak prevalence}

\item{...}{additional arguments are ignored}
}
\description{
Susceptibles at boundary layer (corner/boundary layer approximation)
}
\details{
\deqn{
	x_{\rm in} = 1 + \left(1-\frac{1}{{\mathcal R}_{0}}\right) W_{-1}\left(-\frac{1-x_{\rm f}}{1-\frac{1}{{\mathcal R}_{0}}} e^{-\frac{1-x_{\rm f}}{1-\frac{1}{{\mathcal R}_{0}}}}\left(\frac{y_{\rm max}}{y^*}\right)^{\frac{\varepsilon}{{\mathcal R}_{0}}\frac{1}{1-\frac{1}{{\mathcal R}_{0}}}}\right)
}
}
\seealso{
\code{\link{x_in}}, \code{\link{x_in_cb}}, \code{\link{x_in_exact}}
}
