% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_in.R
\name{x_in_crude}
\alias{x_in_crude}
\title{Susceptibles at boundary layer (crude approximation)}
\usage{
x_in_crude(R0, epsilon, peakprev_fun = NULL, ...)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period (\eqn{1/(\gamma+\mu)}) as a
proportion of mean lifetime (\eqn{1/\mu}; \eqn{\varepsilon =
    0} corresponds to infinite lifetime, i.e., no mortality)}

\item{peakprev_fun}{function of \eqn{{\cal R}_0} and
\eqn{\varepsilon} to use to compute peak prevalence}

\item{...}{additional arguments are ignored}
}
\description{
Susceptibles at boundary layer (crude approximation)
}
\details{
Argument \code{peakprev_fun} is ignored but is present so that the argument
list is the same as for \code{\link{x_in}}.
}
