% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_in.R
\name{x_in_KM}
\alias{x_in_KM}
\title{Crude "Kermack-McKendrick" version of \eqn{x_{\rm in}}}
\usage{
x_in_KM(R0, ...)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{...}{additional arguments are ignored}
}
\description{
FIX: this is the same as \code{\link{x_in_crude}}, but coded
differently.
}
\details{
This is just the case where \eqn{\varepsilon=0},
\deqn{x_{\rm in,KM}({\mathcal R}_0) = x_{\rm in}({\mathcal R}_0, 0)}
}
\seealso{
\code{\link{x_in}}, \code{\link{x_in_crude}}
}
