% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_in.R
\name{Ytilde_1}
\alias{Ytilde_1}
\title{\eqn{{\tilde Y}_1(x_{\rm f})} (vector version)}
\usage{
Ytilde_1(xf, R0, tiny = 1e-12, subdivisions = 1000L, ...)
}
\arguments{
\item{xf}{final size (\eqn{0<x_{\rm f}<1})}

\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{tiny}{amount by which to avoid integration limits}

\item{subdivisions}{the maximum number of subintervals.}

\item{...}{additional arguments to be passed to \code{f}.}
}
\description{
\eqn{{\tilde Y}_1(x_{\rm f})} (vector version)
}
\examples{
R0 <- 2
xf <- 1 - final_size(R0)
Ytilde_1(xf,R0)

}
\seealso{
\code{\link{Ytilde_1_scalar}}
}
