% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_in.R
\name{Ytilde_1_scalar}
\alias{Ytilde_1_scalar}
\title{\eqn{{\tilde Y}_1(x_{\rm f})} (scalar version)}
\usage{
Ytilde_1_scalar(xf, R0, tiny = 1e-12, subdivisions = 1000L, ...)
}
\arguments{
\item{xf}{final size (\eqn{0<x_{\rm f}<1})}

\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{tiny}{amount by which to avoid integration limits}

\item{subdivisions}{the maximum number of subintervals.}

\item{...}{additional arguments to be passed to \code{f}.}
}
\description{
\eqn{{\tilde Y}_1(x_{\rm f})} (scalar version)
}
\details{
\deqn{
\tilde{Y}_{1}(x_{\rm f}) =  \int_{x_{\rm f}}^{1} \left[\left(\frac{1}{{\mathcal R}_{0}t} -1\right) \frac{1-t}{{\mathcal R}_{0}\,t\,Y_{0}(t)} + \frac{1-x_{\rm f}}{{\mathcal R}_{0}x_{\rm f}}\frac{1}{t-x_{\rm f}}\right] dt
}
where
\deqn{
	Y_{0}(x) = 1-x+\frac{1}{{\mathcal R}_{0}}\ln{x}
}
}
\note{
\code{subdivisions} and \code{...} are arguments to
\code{\link[stats]{integrate}}.  The function \code{f} to
which \code{...} is passed is the integrand.
}
\seealso{
\code{\link{Ytilde_1}}
}
